Added delete

+++ database/migrations/2017_05_22_123718_create_users_comment_table.php
@@ -17,7 +17,7 @@ class CreateUsersCommentTable extends Migration
             Schema::create('comments', function (Blueprint $table) {
             $table->increments('id');
             $table->string('body');
-            $table->integer('user_id');
+            $table->BigInteger('user_id');
             $table->integer('commentable_id');
             $table->string('commentable_type')->default('App\\\Article');
             $table->rememberToken();
